removed access to xdg-config, xdg-data, xdg-cache
I see the security reason for flathub blocking access to those directories, but it harms the users who want to easily switch between a system install of Gramps and the flatpak version of Gramps. So this pull request will only add access to ~/.gramps which is what Gramps 5.1 and earlier used.